Ajout du gitignore, détails dans le readme et creation du main
