+++ README.md
@@ -4,3 +4,14 @@
 
 [![codecov](https://codecov.io/gh/Skylabe/TP1-TeamBGs/branch/master/graph/badge.svg)](https://codecov.io/gh/Skylabe/TP1-TeamBGs)
 
+
+## Launch the project
+The following command show how to compile the project using javac:  
+```javac -cp src/main/java/com/esiea/tp4A/domain -d src/main/java/com/esiea/tp4A/domain/bin/ src/main/java/com/esiea/tp4A/domain/impl/*.java```  
+Next, you can lauch the project as follows:  
+```java -cp src/main/java/com/esiea/tp4A/domain/bin MarsRoverImpl```
+
+## Launch the tests
+Test all the project:  
+```mvn test```  
+
