+++ src/main/java/com/esiea/tp4A/domain/Direction.java
@@ -2,3 +2,4 @@ public enum Direction {
     
     NORTH, EAST, SOUTH, WEST;
 }
+
