+++ src/main/java/com/esiea/tp4A/domain/MarsRover.java
@@ -18,3 +18,4 @@ public interface MarsRover {
         return Position.of(0, 0, Direction.NORTH);
     }
 }
+
