+++ src/main/java/com/esiea/tp4A/domain/PlanetMap.java
@@ -4,3 +4,4 @@ public interface PlanetMap {
     
     Set<Position> obstaclePositions();
 }
+
