+++ src/main/java/com/esiea/tp4A/domain/Position.java
@@ -37,3 +37,4 @@ public interface Position {
         }
     }
 }
+
