+++ src/main/java/com/esiea/tp4A/domain/bin/.gitignore
@@ -1,0 +1,3 @@
+# Compiled class file
+*.class
+
