+++ src/test/java/com/esiea/tp4A/domain/impl/PlanetMapImplTest.java
@@ -1,0 +1,36 @@
+package com.esiea.tp4A.domain.impl;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+
+class PlanetMapImplTest{
+
+    public PlanetMapImplTest() {
+    // implementer taille map
+        
+    }
+    
+//    @ParameterizedTest
+//    final void testobstaclePositions(){
+//        Set<Position> mapObstacles = new HashSet<Position>();
+//        for (int x = 10; x < 20; x++) {
+//            for (int y = 10; y < 20; y++)
+//                mapObstacles.add(Position.of(x, y, Direction.NORTH));
+//        }
+//        return mapObstacles;
+//    }
+//    
+    
+//    @ParameterizedTest
+//    public static void main(String[] args) {
+//        PlanetMapImpl p = new PlanetMapImpl();
+//        Set<Position> a = p.obstaclePositions();
+//        /*for (Position pos : a) {
+//            System.out.println(a.getClass().getName());
+//        }*/
+//        System.out.println(a.getClass().getName());
+//    }
+}
+
